+++ Utilities/Utilities.d.ts
@@ -1,10 +1,17 @@
 declare namespace Utilities { }
 
+declare class Path {
+    public ModuleDir: string;
+    public ModuleMain: string;
+    /**
+     * Convierte una ruta relativa en una completa.
+     * @param Path La ruta relativa.
+     */
+    Relative(Path: string): string;
+}
+
 declare class Utilities {
-    public static Path: {
-        ModuleDir: string,
-        ModuleMain: string
-    }
+    public static Path: Path
 };
 
 export default Utilities;